Rename: stay type filter 관련 컴포넌트 filter 폴더로 분리
