+++ src/components/layout/HeaderContainer.jsx
@@ -4,7 +4,7 @@ import { throttle } from 'lodash';
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
 import { PATH_URL } from '../../constants';
-import Filter from '../header/Filter';
+import Filter from '../header/filter/Filter';
 
 const HeaderContainer = () => {
   const [isscrolltop, setIsScrollTop] = useState(window.scrollY < 40);
